optpython/requirements.txt
astroid==2.1.0
colorama==0.4.1
et-xmlfile==1.0.1
isort==4.3.4
jdcal==1.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.15.4
openpyxl==2.5.12
pandas==0.23.4
py-lets-be-rational==1.0.1
py-vollib==1.0.1
pylint==2.2.2
PyQt5==5.11.3
PyQt5-sip==4.19.13
python-dateutil==2.7.5
pytz==2018.7
QuantLib-Python==1.14
scipy==1.2.0
simplejson==3.16.0
six==1.12.0
wrapt==1.10.11
